Add helper function to load and parse in one shot.
The lack of monads in C++ unfortunately inflates the amount of helper functions
to chain together monadic calls, *sigh*. Here's a new helper for the sequence
of loading and then parsing file contents: load_and_parse<T>(path p). This
function (i) first loads the file contents, and (ii) then attempt to parse it
as usual with to<T>(string const& str, ...).