Streamline and optimize bitmap implementation.
The notion of "append" for a bitstream was ill-defined: it doesn't make
sense to append bits to a bitmap because it operates at the granularity
of arithmetic values. We (ab)used the append operation to artifificially
increase the number of rows in the bitmap, so that future push_back
operations work with the value corresponding to the correct event. Now
we have a new explicit function "stretch" which performs a virtual
append: it mereley increases the number of rows. A push_back was already
lazy in the past and we leverage push this now even further by
completely obviating the need for *any* bitstream operation during
stretching. Previously we appended 0 bits to all bitstreams of the
bitmap. So this change should come with a significant performance
increase and bitmap insertion rate.

Furthermore, one can now push_back the same value multiple times.
bitmap::push_back has a second parameter which specifices the number of
times a value should be appended. We plan to leverage this in the future
for batch-append operations, e.g., if the value 42 occurrs N times
consecutively in a stream of events, we do not want to invoke N times
push_back(42, 1) but rather push_back(42, N).