Add pseudo-realtime mode for PCAP source.
A common problem with trace analysis is that the reading application
slurps in all packets as fast as it can, which ignores the inter-arrival
time of packets as on the wire. When evaluating the performance of a
system, this becomes a confounding factor.

To this end we have now added "pseudo-realtime" mode when reading
packets from a network trace. The PCAP source now sleeps for time t/c
between two consecutive packets spaced apart by time t > 0, and some
constant c > 0 representing the speedup or acceleration factor.

For example, to import a trace in real time, one would do this:

    vast -I pcap --import.pcap-pseudo-realtime 1 < trace.pcap

Note that the PCAP source should run in a separate thread, i.e., spawned
in detached mode. Otherwise it causes a worker thread to sleep and waste
cycles. But since the PCAP source does file/network I/O, one would not
want to hook it into the scheduler anyway.

Closes #62. Relates #51.