Enhance query expression normalization.
VAST performs the following normalization steps:

  - In addition to hoisting single-element conjunctions/disjunctions,
    hoisting also applies to nested operations. E.g., the query

        A && (B && C)

    gets normalized to simply

        A && B && C

  - Negations are pushed down to the predicate level and
    double-negations eliminated. E.g.,

        !! A && ! (B || C )

    becomes

        A && && ! B && ! C

    This is equivalent to creating transforming the expression into
    *negation normal form*

The negation normal form is a crucial for correctness of query results.
If we did not have it, we would be able to create non-monotic streams of
hits. Consider Q = ! (A && B). As the hits for A and B trickle in Q
decreases in number of hits. This is problematic because we would have
block until Q has completed in its interity. In other words, negation
normal form allows for asynchronous computation of hits.