Make task exit message customizable.
When a task is done, it sends a DONE message to its supervisors which
had the static type <done_atom, time::duration>. But when we have
multiple tasks running within the same actor, our handler becomes
cluttered with checks to figure out which task this DONE message
corresponded to. Therefore we made the tasks more flexible by making it
possible to specify the DONE message upon construction. A task now has a
variadic construct and one can create tasks with many different types of
the form <done_atom, time::duration, Ts...>. This allows an actor
hosting multiple tasks to simply specify different handler types to
associate completion code with the right task.