Fix bug in flow-control signal propagation.
In this case our own abstractions for flow control shot us in the foot.
The use of become(foo) switched an actor's behavior, but 'foo' here
should have been augmented by the system handlers. Because we used the
naked behavior instead of the augmented, we lost a few important
handlers, such as the one for DOWN. What this really means is that we
need a better mechanism for behavior composition. A very first step
would be actor-framework/actor-framework#227. With this functionality in
place, we no longer have to provide hooks in the form of virtual
functions that derived actors need to fill in. Instead, derived classes
can then define their match expressions directly as part of their
behavior. We'd still need something like become(augment(foo)) or
custom_become(foo) when switching behaviors, but at least we'd have less
inversion of control and could reason more locally about the code.

Further changes:

  - We also propagate the flow control messages now all the way up to
    the source, and disburdened IMPORTER from handling them.

  - Remove sleep from unit tests and insert ping/pong barriers to ensure
    we're not terminating actors too early.