Improve shutdown performance of partitions.
It's not necessary to flush partitions if nothing has changed and
indexers flush on exit anyway. So now we only have to ensure that all
state has been saved when an actor exits and then can switch to a
terminating modus where we delay quitting until all of our children have
quit.