Change behavior of how to find CAF.
When setting --with-caf=X we allow now that X is either an installation prefix
or a CAF *build* directory. In the latter case, we assume a standard
installation with the build directory in the source tree, and attempt to find
the CAF headers there. This makes it possible to have multiple build
directories in the same source tree and select the specific libraries
accordingly while sharing the same include path to the headers.