Add man page generation infrastructure.
We use md2man [1] to write man pages in Markdown and generate a roff
version from it. When CMake finds md2man, it adds the following new
targets:

  (1) make man: generates the man page in the build directory.

  (2) make man-view: invokes 'make man' and then launches the man viewer
      to open the freshly generated man page.

  (3) make man-replace: replaces the current man page in the source
      tree. VAST always ships with the latest version of a man page so
      that users don't need to install md2man just to get the most
      recent version. This command replaces a freshly generated man page
      with the version in the source tree.

The current contents document a future version of VAST, where we're
aiming towards with the development. It does not reflect the current
architecture.

[1] https://github.com/sunaku/md2man