Move towards a distributed container architecture
This commit introduces a fundamental architecture change. The vast
executable can now either be in server mode, where it spawns a single
NODE actor which acts as container for other actors, or it can ran in
client mode, where it connects to NODE via a TCP connection and sends it
a command. See the new man page in doc/vast.1.md for details.

In general, NODES can peer with each other, which means they share
global state. This is implemented in terms of the new distributed
key-value store. When creating multi-machine topologies, this becomes
important: the store tracks which actors exist on which machines and
enforces a global unique namespace, e.g., two archive actors cannot have
the same name.

Users now interact with VAST as follows:

        vast start -c   # spawns a VAST core
        vast spawn source bro -r /path/to file
        vast connect source importer
        vast send source run

While these are now 4 distinct commands, they offer much greate
flexibility in terms of designing topologies. In the future, we will
have "meta" commands for command cases such as import and export.