Add utilities to test file descriptor sending
These new toy examples showcase how one can send file descriptors via
UNIX domain sockets. In particular, the server accepts a client
connection, attempts to read a descriptor from the socket, then writes
to it, and then closes the connection. The client connects to the
server, sends the file descriptor for standard output (= 1) over the
socket, and waits for the server to write data into it. Ultimately, on
the console the client sees the data written into the descriptor by the
server.