Use more consistent base class for sources
Renaming the class source::synchronous to source::base is consistent
with the existing sink::base class. The "synchronous" aspect is a
left-over from a time where we considered both synchronous and
asynchronous source types. If we end up with such source types in the
future, we can still make that distinction.