Make sources produce events rather than chunks
This change further streamlines and simplifies the ingestion process.
A SOURCE now simply produce sequences of events. In a cluster-in-box
setup where all actors run in the same NODE, this is more efficient,
because it avoid the extra decompression step at INDEX. Even in the
distributed case, it disburdens the CPU-intensive INDEX (no more chunk
decompression) and moves the computational load to ARCHIVE (compress
events into chunks).