Add support for literal parsers
By making the DSL operators free function and using SFINAE to constrain
the overload set, we can now convert binary expression of the form

        LHS op RHS

into a parser if both LHS and RHS are parsers, or if either side is
convertible to one. It is possible to create parsers with unused
attributes from character and string literals, as well as from
arithmetic types (except bool).

For example, the following statement would create a parser for the
string "<foo>":

        auto p = '<' >> string_parser{"foo"} >> '>';

The parser has the synthesized attribute std::string, because the
attributes of the two literals are ignored.