Fix issue with IPv6 address parser
Until we decode each byte manually, we rely on inet_pton. However, this
function requires a NUL-terminated string and does not work when the
address string has trailing characters. That is, it doesn't parse out an
address of an arbitrary string (although it should). This commit works
around this issue.