Update Docker image on every build
By putting the ADD instruction before the CAF dependency setup, we
ensure that each build fetches the most recent version of the develop
branch in CAF. Otherwise the build cache may use a stale CAF version and
try to build VAST against it.

The correct fix would be creating a CAF Docker image.