Change ID management and peering behavior
Previously, all nodes with IMPORTERs shared a single IDENTIFIER located
on one of the nodes. This is a central point of failure and not
desirable. The architectural change is that we have now a distributed
key-value store with linearizable semantics. Eventually, the consensus
module will use Raft. (For now we use a *very* stripped down version with
manual leader election and two-phase commit for log replication.) The
key-value store keeps track of the global event ID. Whenever an
IDENTIFIER needs more IDs, it will perform an ADD X request to the
store, which will increment the counter by X and return the old and new
value deliminating the requested IDs.

IMPORTERs now ask the local IDENTIFIER until it cannot supply enough
IDs. Then IDENTIFIER makes a global request for more IDs. If IMPORTER
performs an exhaustive request that depletes all local IDs, IDENTIFIER
will double its request size. This exponential growth (akin to TCP slow
start) quickly finds the "right" batch size for the event rate at
IMPORTER, so that it doesn't have to bother the global key-value store
often. Also, if an IDENTIFIER exausts its IDs twice within 10 seconds,
it will double the request size as well.

Until we our distributed consensus module supports leader election,
peering semantics are now asymmetric: When node A requests to peer with
node B, it will always assume B is a leader and attempt to follow.

Other implementation details:

  - Use vast::key to construct hierarchical keys. Previously we used a
    std::string and hard-coded the separator '/', which is prone to
    errors.

  - Equip key-value store with an ADD operation to increment an existing
    or (unit) value. This currently works only on arithmetic types and
    strings.

  - Some keys in the key values store can be marked persistent. They
    survive node restarts and get written to disk after each successful
    replication from leader to its followers.

  - The shutdown process of NODE now has an order: it begins with
    terminating IMPORTERs after they have processed their queues. This
    ensures that events submitted from outside sources will not get
    lost, which can happen when downstream ARCHIVE/INDEX actors
    terminate before IMPORTERs.

Resolves #81.