Improve support for local node operations
For testing or scenarios where one "just wants to get going," it is now
possible to run all of VAST in one shot when running the vast
executable. If one doesn't specify and endpoint with -e, the vast
executable now assumes a local setup and spins up a node with core
actors before processing the command line. This comes in handy with the
"import" and "export" commands. For example,

        zcat *.log | vast import bro

now just imports Bro logs into a node running in the same executable.
Similarly,

        vast export ascii -h 'orig_h in 10.0.0.0/8'

now spins up a local node prior to issuing a query. This won't change
the existing client/server semantics:

        vastd -l 5 -f
        vast -e 127.0.0.1 import bro < conn.log

The above still spawns a node listening on :42000 and allows remote
queries.