Improve endpoint handling
The new option -n makes it explicit when VAST starts a NODE. Previously,
the "import" and "export" commands always spun up a local NODE, which is
not a consistant command line interface and can lead to unexpected
filesystem races when a daemon happens to runs as well.

The "start" command implies -n by definition, so there's no ambiguity
here.