Implement HTTP routing and query execution (WIP)
The HTTP layer architecture begins with a HTTP broker, which forks one
HTTP worker per incoming connection. Each worker then parsers the
request and sends it off to MEDIATOR, an actor which takes the
information from the request and translates it into VAST commands, e.g.,
like spawning a new EXPORTER. MEDIATOR maintains a set of SHEPHERDs,
actors which accommodate a set of EXPORTERs and acts as their sink. When
new results arrive, SHEPHERD converts them into JSON and creates a
response which goes back to the corresponding HTTP worker.

Related changes:

  - EXPORTERs no longer terminate automatically after the query
    terminates. This allows for inspecting the hits again and use the
    EXPORTER to seek in the result set in order to extract previously
    delivered results.

    FIXME: this change broke 'vast export'. Now the standard form of
    querying needs to be ported before signing this commit off.

  - There is a new exit reason 'timeout'.