Streamline interaction of types and data
This large set of changes concerns the relationship between types and
data. Along the way, we factored the unit tests of the respective data
types and improved the implementation of the printable concept, given
its new enhancements.

Moreover, this commit brings to the table Howard Hinnant's chrono_io
library to format durations uniformly.