Revamp bitmap implementation
This commit restructures key parts of VAST's building block: the
efficent implementation of bitmaps. We no longer user the term
"bitstream" to refer what the community generally understands as a
bitmap. What we used to call "bitmap" in VAST is technically a "bitmap
index," and will address this terminology discrepancy in a future
commit. For now, bitstreams are gone and replaced by bitmaps.

In addition, the class vast::bitmap offers a type-erased bitmap
implemention, in cases where the concrete compression algorithm doesn't
matter. For now, we can only select between EWAH and no compression at
all, but this will change in the future. The idea is that bitmap indexes
only operate on the type-erased structures, and that we can construct
generic algorithms that dispatch based on the concrete bitmap types
(e.g., provide an optimized overload for vast::bitwise_and(T, U) for two
different bitmap types T and U).

In general, the revamped bitmap API puts more emphasis on algorithms as
free functions, in contrast to emphasizing data structures. The encoded
bitmaps only provide a range that enables bit-sequence iteration, and
the algorithms then make use of these ranges.

Finally, we've moved the bit-range-based iteration capability of
vast::bitvector into vast::null_bitmap, because std::vector<bool> (which
vast::bitvector emulates in platform-indpendent way) neither offers
block-level algorithms.