Implement bitwise evaluation for many bitmaps
Wu et. al describe in Section 5 in their 2004 paper titled On the
Performance of Bitmap Indices for High-Cardinality Attributes various
strategies to  evaluate bitwise logical operation on multiple bitmaps.

This commit implements "Option 3" as default method. Once the
performance of these lookups turns out to be insufficient, we will
switch to an adaptive scheme, as described in the paper.

However, we have to re-evaluate the numbers presented before doing so.
The reason is that the underlying implementation (fastbit) appears to
contain a bug that may have affected the numbers in the paper:

    typedef std::pair<ibis::bitvector*, bool> _elem;
    // put all bitmaps in a priority queue
    std::priority_queue<_elem> que;

This snippet stems from index.cpp(5619) of the most recent release of
fastbit (as of this commit), where the algorithm defines a priority
queue. The problem is that the priority_queue needs a custom comparator
in order to create a min-heap with meaningful semantics. Here, std::pair
uses element-wise operator<, which results in comparison of memory
addresses, instead of the actual bitmap size.