Make bitmap evaluation compression-agnostic
Since our bitmap evaluation algorithms operates with a notion of bit
runs, the concrete bitmap compression algorithms does not matter. For
example, we can now perform bitwise operations where one operand is a
compressed bitmap and the other uncompressed. The type of the result is
determined by a type trait, akin to std::common_type.