Implement set/vector index
Also provide an internal utility for polymorphic serialization. This is
a pretty hacky situation, but since there's no support for polymorphic
types on the CAF side, we have to bite the bullet. In this case, this
means we emulate a dynamic_cast by dispatching on a type instance, which
boils down to dispatching on a variant.