Do synchronize buffer with underlying streambuf
The problem is that we don't know whether we're in reading or in writing
mode. It would be fatal to write to an immutable sequence. Therefore,
users have to call pubsync() on the stream buffer directly.

In the future, we could perhaps figure out, based on the get area
semantics, whether we're in reading or writing mode.