Change handling of nil values
VAST now excludes nil values from lookups. The consequence is that, for
example, !(x == 42) yields a different result from x != 42. This may
seem counter-intuitive first, but it is consistent. While it may seem
natural for x != 42 to include nil values (because they are not equal to
the value 42), for <, <=, >=, > it becomes less clear: should nil be
less or great than any other value in the domain? Since nil is not part
of the domain under which the operator is well-defined, lookups no
longer contain nil results. If nil values are needed, it requires a
separate clause, e.g., x != 42 || x == nil.