Build CAF inside VAST Jenkins workspace
The reason is that a specific build config of VAST requires specific
build parameters for CAF as well. Since the two projects have different
build matrices, they are a difficult to combine, even though they are no
the same machine. In particular, we encountered this issue [1]:

    We know about at least one kind of false positive container overflow
    reports. If a part of the application is built with asan and another
    part is not instrumented, and both parts use e.g. instrumented
    std::vector, asan may report non-existent container overflow.

Thus, we need to synchronize at least the ASAN flags for CAF and VAST to
produce successful builds.

[1] https://github.com/google/sanitizers/wiki/AddressSanitizerContainerOverflow