Add new view infrastructure
A view is an abstraction over a structure that exists in serialized
form, where deserialization/conversion is performed lazily at access
time. A view always points to a reference-counted chunk of memory. That
is, the last view destroys the corresponding chunk. For chunks that do
not own their memory, it's the user's responsibility to ensure that
views don't outlive the chunk.

The view API involves *packing* data to chunks and *unpacking* data from
chunks.