Add packers, overlays, and unpakers
A packer serializes sequences of elements into a contiguous chunk of
memory. One can attach an overlay to this chunk to get a lazy sequence
view that provides random access into the chunk. This enables selective
deserialization of an element i, as opposed to having to deserialize all
elements j where j < i.