Rewrite MRT parser with combinators
This commit takes Philipp Meyer's work as a baseline for the MRT
protocol analysis. We rewrote the parsing logic in terms of parser
combinators and VAST's parseable concept.

Moreover, we now read the input one MRT record at a time to support
incremental reading through a (conceptually unbounded) file stream.