Consolidate append API of bitmap indexes
The push_back function is now called append, to avoid confusion with a
container from the standard library. This refactoring commit also
separates the notion of "skipping" and appending, which was previously
muddled into a single function. Skipping a value is a cheap operation,
because it doesn't require encoding of a value. It merely "stretches"
the bitmap index so that future append operations start at a specific
offset. It's up to the user of a bitmap index to mask out skipped
values. (For example, that's what value_index does.)