Port the IMPORTER to table slizes
Open to discussion: The maximum slice size is currently fixed. All
sources that produce fewer items per slice currently cause gaps in the
ID space. Keeping a per-source track on slice sizes makes credit
management a lot more complicated, so we should only go down that route
if we measure a need to do so.