Consolidate debug builds in Jenkins
Having debug builds with multiple configurations (with and without ASAN,
with and without Logging, etc.) is generally desirable. However, we
currently overbook the available workers and thus have to wait for one
hour per build. This change makes sure no builds land in the queue (if
all workers are available at the start).