Fix failing unit test
This test was wrong in the first place. A query N.a.b.c where N is a
type name an a.b.c a sequence of nested record fields shouldn't be part
of the query language in this form because it's confusing for users.
(In fact, users already can specify type names via the &type attribute.)
Consider this type:

  type T = record { T: int }

A query T.T does not only look unclear, but it further can be
misinterpreted. Is it the suffix any other record named U where T != U?

Instead, what we want is a unique symbol, say #, that distinguishes
record suffixes from actual type names. The expression

  #T.T

would then be interpreted unambiguously with #T referencing the type
name and .T a dereferenced record field.