Remove optimized mmapbuf::resize() method
After further testing, the approach of mapping an adjacent
memory region turned out to be buggy on darwin as well, occasionally
triggering a SIGBUS. A (non-rigorous) performance comparison did not
reveal a significant runtime difference between this method and the
de-pessimized (see below) full remap method.

The full remap method contained an unnecessary memcpy from the original
to the new mapping, which has been removed.