Remove non-standard __attribute__((packed))
GCC rightfully rejects forming a reference to members in a packed
struct. Pointers must follow the platform-dependent alignment, which
the compiler can no longer guarantee when using the packed attribute.
However, we don't need the attribute in the first place. All members
occupy 4, 8, or 16 bytes. This means the in-memory representation is
always going to be aligned on any platform we care about.

Instead of using the non-standard, alignment-breaking attribute, we can
simply double-check our assumption with a simple static_assert here.