Avoid name resolution by default
There seems to be a mysterious issue going on where "localhost" either
no longer resolves to "127.0.0.1" (not the case on my machine) or some
low level socket functions do weird stuff. In any case, when I use
broker::endpoint::listen("localhost"), the process doesn't listen on
127.0.0.1.

As a workaround, 127.0.0.1 is now the hard-coded default.