Factor get-bitmap-then-fill pattern into function
The coders have many places where they get a bitmap with a specific
index and then fill it to the current size to make sure it's safe to
modify further. This access pattern is modeled by the new `bitmap_at`
function. Further, the index-based access in range_coder::decode was
wrong for operator `<` and `>=` on index 0. This is fixed by returning
default values in those cases.