Fix build on FreeBSD
This is an interesting issue. When adding a value of type
std::chrono::seconds to (caf::)timestamp, we don't get a caf::timestamp
back!

  template <class C, class D1, class R2, class P2>
  constexpr time_point<C, typename std::common_type<D1, duration<R2,P2>>::type>
  operator+(const time_point<C, D1>& pt, const duration<R2,P2>& d);

The culprit is this type:

  typename std::common_type<D1, duration<R2, P2>>::type

If it would evaluate to D1, we'd be fine. But it takes duration<R2, P2>.
Now, what does the standard say for std::chrono::seconds?

  duration</*signed integer type of at least 35 bits*/>

In other words: undefined.