Fix pesky reference-to-temporary bug
In the statement

  auto str = "foo";

the lvalue str decays to const char* and not a char[N]. We only had a
char[N] overload that creates a view<string>. Once we had a const char*,
the overload with const data& got picked. And because data has
non-explicit constructor, make_view created a temporary data object and
return a view to it.