Restructure construction of table slices
This commit makes several changes to how table slices are constructed.

Most notably, We no longer use an actor system to store the factory for
table slices. Instead, this is now global state. The rational for this
change is that the runtime settings map in an actory system has a flat
key space for all type of settings. Because table slice may be
constructed in multiple ways with multiple factories, relying on an
actor system would make the factory management cumbersome. Even though I
don't like global state as well, this mechanism of extension is far less
complex that invovling the actor system.

Now, after implementing two factories and managing them separately, I
found a simpler solution that only involves a single factory: make it
possible to default-construct a table slice through a factory and then
provide a virtual function load(chunk_ptr) to initialize the chunk.
The default implementation of 'load' deserializes from the chunk using the
deserialize code. That is, 'load' is just an optimization for table
slice implementation that are implemented like a view.

This change required default-construction of table slices, which
entailed several more changes, including changing the implementation of
the matrix table slice. While it was technically appealing to have an
implementation that performs a single round-trip to the allocator, this
turned out never to be the bottleneck, but it prevented
default-construction. The API changes now add one extra allocation for
construction the table slice before population it with data as before.

Upleveling, there now exists a new requirement that a table slice in
serialized form *always* has to begin with its ID (in network-byte
order). This ensures that (de)serialization and loading from a chunk can
use the same factory.