Switch to flat record parsing
Moreover, we no longer have an additional timestamp column prepended to
every table slice. It's just the input data now.

This commit also tweaks the schema of the readers by adding a &time
attribute to indicate that the column represents the event timestamp.
Now, the meta index uses this attribute to resolve &time queries. If a
reader has a &time attribute in one of its fields, it will be used a
event timestamp.