Fix two nasty bugs in the table indexer
The first bug was that we used the wrong directory to identify the file
containing the row IDs. All table indexers in a partition used the same
file, instead of using one per layout.

The second bug was a double-flush that ocurred in the destructor of a
moved-from table indexer. In that object, a flush should clearly not
happen.