Set artifact path at test executable build time
This change pushes the injection of the artifacts path
from libvast_test down to its dependees. Before this
the path was hardcoded in the events fixture, so if the
installation was moved, it wouldn't resolve any more.
Now, the path is injected later when the test executable
is built.
This also removes the autogeneration of the path header
and puts the path variables for each artifact into a
deeper namespace (e.g artifacts::logs::zeek etc).