Always load default schemas
This change lets VAST load its schemas from a path relative to the
executed binary, instead of the installation prefix. With that
there is no more use case to overwrite the default paths, so
the `--schema-paths` option won't do that any more.
Note that static executables are only supported on linux right now.