Support new Data Predicates
A data predicate is a short-hand for a predicate of the following form:

  - LHS: type extractor
  - op:  equal
  - RHS: data

When the input only consists of data, we can assume that the user wants
to say "give me everything that equals this". That's what type
extractors are for. Without being more specific, we can also assume that
equality comparison is the desired operation.