Add schema file based on ra(1) manpage
The -s section in the ra(1) man page contain a long list of options to
be passed to the ra. The list in the man page served as basis for this
commit. Our script to extract labels provided also some extra fields
that are not mentioned in the man page.