Set socket connection timeout
It seems that on Cirrus `create_connection` frequently times out on the first try but succeeds on the second try. The default timeout is 60 seconds which seems too much for just checking availability of a port. This change set default timeout to 1 second which helps mitigate the issue of frequent socket timeouts on the first attempt.