Improve value index performance and storage
The value index keeps two extra bitmaps independent of its
implementation: a null bitmap and a mask bitmap. The null bitmap is for
all nil values and the mask value used to be for all IDs that exist in
the index *and* the nil values.

This change makes the two bitmaps disjoint. The result is not only less
bits overall, but also less bit-wise operations when performing a
lookup, because the null bitmap no longer needs to be NANDed out.