Fix race in index lookup
The lookup protocol that the index exposes is as follows:

+--------+ -- expression --> +-------+
| Client |                   | Index |
+--------+ <- lookup meta -- +-------+
      ^                          |launches
      |                          v
      \---- Ids ----------- +-----------+
                            | Evaluator |
                            +-----------+

The message sent from the Index to the Client is the _response_,
the message form the Evaluator to the Client is the _result_. The
actor model cannot give us any guarantees on the order of message
delivery as soon as an indirection is involved.

The problem is that the index used to start the Evaluator(s)
before inserting the response message into the receiving queue of
the Client. Thus it was possible that the Evaluator placed its
messages into the same queue at an earlier time, violating the
requirement of the client that the response shall be received
before the results.

This commit changes the order of response delivery and lauch of
the Evaluator(s), thereby fixing the bug.