Fix CMake project scaffold
Any user supplied value for `VAST_VERSION` was overwritten to the
empty string at the call of `project(VAST)`. This change renames the
variable in question to `VAST_VERSION_TAG` as a workaround.

The implicit extraction of `VAST_VERSION_TAG` is now done before the
call to `project` to make this non-obvious behavior more visible.