Make variants in test comparison macros explicit
The CHECK_EQUAL() macro has a lot of additional behavior inherited from
caf, most significantly the automatic unboxing of variant types. This
leads to `CHECK(x == y) != CHECK_EQUAL(x, y)` in many cases.

With this commit, the `CHECK_*` and `REQUIRE_*` families of macros
will only perform the comparisons that they advertise, and the
programmer has to explicitly opt-in to the additional behaviour
using a different set of macros.