Do not use shallow commits when fetching tags
When running the fetch step in our github action workflow

    git fetch --depth=1 origin +refs/tags/*:refs/tags/*

we are *replacing* all existing commits that are tagged on the
remote with shallow commits.

For some reason that is not yet understood, this breaks the
output of `git describe` on the vast repository, calculating
an incorrect number of commits since the last tag.