Remove assertion from expression tailoring
Removed an assertion that checked an expression could
not become 'none_t' after running the type resolver on
it.

Since the type resolver is using 'none_t' as a return value
to signal that it was not able to resolve a type, this
condition should not be an error, let alone an assertion
failure.

For example, an expression like 'foo.bar = "baz"' would always
return 'none' if the type to be matched against was not called
foo or had no field called bar.

In addition, also adjust the call sites of 'tailor()' to not
interpret 'none_t' as "this checker was not yet created".