Fix file-verbosity settings
This inverts the logic to suppress the to file logging of client
commands. The old algorithm overwrites the verbosity setting with
the default value after it was parsed, if the command is either
`vast start` or `vast -N ...`. This means that user specified
levels from the config file or the command line would have no
effect.

The new method does the opposite: Once we know the command does
not start a node the file logging is turned off.